netlist: Add a function to get a unique unnamed connection name.
Previously, the name of an unnamed connection depended on the
coordinate of the first object of the connection from its object
list.  A new function has been introduced that chooses the same
connection point, the coord of which is used then in forming the
name of the connection independently of the order of objects in
the object list of the connection.

The algorithm is simple: of all connectable objects, get the joint
point coordinate which has the least X, and then, if there are
several of them having the same X, get one which has the least Y.

The joint point coordinate is any end of a net for nets and the
connectable point of a pin for pins.

The tests have been amended accordingly.