lepton: Add syntax rule to check if object is attribute.
The syntax rule checks if an object is attribute.  It raises an
error with the key 'wrong-type-arg if it is not text, and
'attribute-format if the object is not attribute text, that is, it
cannot be parsed as attribute.  Otherwise the rule returns
attribute name and value as two return values.