manual: Suppress building DVI docs on 'make distcheck'.
This was an unnecessary step which also required building EPS
files that are larger than PDFs for same images.

Apart from that, another issue with producing EPS files is that
lepton-cli itself produces them blurred compared with PDFs, so I
used two steps to provide 'sharp' pictures: first exported
schematics/symbols to PDF with lepton-cli, and then processed them
with pdftops to get the EPS files.  The latter step is now
dropped.