contrib: Replace gschupdate & gsymupdate with an utility in Scheme.
A new utility, lepton-update, has been introduced to replace the
two obsolete scripts.  It works on both symbols and schematics and
does the job a bit better.  It not only replaces the text strings
of attributes but also updates text and pin objects.  Extension of
backup files has been changed from '.old' to '.bak'.