its[config]: Use lepton-export instead of lepton-cli in 'sudo' tests.
This fixes the 'sudo' tests broken due to new lepton-cli code.  As
lepton-cli uses execle() and starts lepton-export as a separate
process, we cannot catch its stdout and stderr output using
command-values().  It would not only require exporting the current
environment to it (e.g. by using system() instead of system*()),
but also would need a special wrapper to catch the stderr and
stdout of the spawned process which runs under superuser.  Taking
into account such a complexity, and because we know from other
tests that lepton-cli always starts lepton-export with right
arguments, it's much simpler just to launch the latter program
itself and test it under sudo.  In such a case, it is sufficient
to use the functions system*() and command-values() the same way
they are used in other tests.