its[shell]: Skip tests that may break on *BSD systems.
It was reported by @graahnul-grom that some tests break on his
system while they pass without errors on Debian and its
derivatives.  Most of issues is probably due to rearranging of
'-l', '-s', and '-c' options by 'lepton shell' on that version
of *BSD before processing any options.  This commit attempts to
mitigate the issue by skipping the tests for now.