build: Retrieve real names of existing libraries with ldconfig.
A new configure stage variable, USE_LDCONFIG, has been introduced
to aid in building Lepton on systems that can miss C headers due
to separation of main and development files ('-dev' or '-devel'
packages).  The main example is Debian and its derivatives.

The fix solves the issue of dlopening shared libraries by Guile on
such systems if the development packages of the libraries Lepton
depends on have not been installed.  In such cases, there is no
safe and portable way to find the necessary libs.  Using C build
utilities like 'ldd' or 'objdump' is not portable.  The manual of
the 'make' program mentiones 'ldconfig' as a utility which can be
used in Makefile's in portable way using '$(LDCONFIG)' syntax and
that utility would be most convenient to use in such cases.
Unfortunately, despite 'ldconfig' is internally used by libtool,
the LDCONFIG variable is not available for 'make' even on Debian.
The solution is specifying 'ldconfig' as a ./configure variable:

./configure USE_LDCONFIG=/sbin/ldconfig ...

After running this command, the necessary library names will be
hard-coded in the appropriate Guile code and thus be available
even when the user has no development packages for them installed.