lepton: Fix regression emerged in commit 485145f.
The code fixes a regression emerged after rewriting string->page()
in Scheme due to using wrong function for string length
calculation.  The contents of UTF-8 characters from extended,
non-ASCII set may take more than 1 byte.  Therefore Scheme
function string-length() cannot be used for byte length
calculation of a string buffer used in C code, namely in
o_read_buffer().  Fortunately, it is not necessary to provide that
length for the latter function if the string processed is
NULL-terminated.  It is enough to set that argument to -1 in such
a case.