gui: Get rid of 'enabled' field in LeptonColor.
The field is really not necessary since its only usage was
disabling some colors in Scheme.  The structure LeptonColor has
been simplified and now the value of color alpha channel is used
to decide whether to draw color or not.  Indeed, when it is equal
to zero, the color is anyways transparent and not drawn.