schematic: Restore behaviour of the complex-place-list-changed-hook().
In its new reincarnation, the hook works with a placement list
containing an inserted component and its toplevel (floating)
attributes.  The auto-place function in the config example,
however, was able to work only with single component objects.
Both the config example and the auto-place function have been
amended so their usage is now correct and the hook name is
unambigous.