gui: Relax requirements to key modifiers processing.
Instead of applying GDK_MODIFIER_MASK to resulted key modifiers on
creation of a GschemKey object, it is now done during event
processing.  It is more sloppy though hopefully, in another use
case of the function g_make_key_struct() it is not necessary since
it gets its arguments from the Gtk string parse function
gtk_accelerator_parse() which should do a decent job.  This should
ease translation of the code to Scheme.