schematic: Merge three scripts and some other code into a module.
The scripts are:
- auto-place-attribs.scm
- auto-place-netname.scm
- default-attrib-positions.scm

They all were used to provide autoplacement functionality in the
file conf/schematic/deprecated.scm.  So the name of the new module
is (schematic autoplace), for brevity.

Moving two previously available variables,
'default-position-of-text-attributes' and
'autoplace-attributes-grid', to the code is necessary to not break
the module.  Some better mechanism is desirable to make the like
settings available for end users.