gui: Fix quirks in GUI attribute insertion behaviour.
In the case when there were no selected and visible objects on the
lepton-schematic page, insertion of an attribute using the Attrib
Edit dialog could lead to some strange result due to arbitrary,
uninitialized, values of proposed coords for it.  The logic has
been changed that reasonable defaults are set in all of the
following cases:

- when there is no object selected,
- when there are no visible objects on the page at all,
- when the mouse pointer in one of the two above cases is over the
  canvas or not,
- when there is no canvas at all.

Now, if there is no object visible on the canvas, and the mouse
cursor did not point to the canvas (was out of the program window
or over the menus) during insertion of an attribute, the program
places a new attribute in the center of the canvas.

The latter case of the above does not obviously apply to the GUI
behaviour.  It adds a last resort coordinate value,
specifically (0 . 0), for cases when the corresponding C functions
could be used as foreign in Scheme code.

The commit also fixes the issue with overriding of coordinate
values set by related Scheme hooks in cases when the action is
launched by hotkey when the mouse cursor is over the canvas.

The Scheme function add-attrib!() has been fixed as well so it
always sets the zero value for coordinates as they are missing in
the set of its arguments, to ensure its proper behaviour in
scripts.