gui: Disable GDK_SMOOTH_SCROLL_MASK if cursor warping is enabled.
The function gdk_device_warp() seems to be incompatible with the
smooth scroll events on zooming.  When it is used, event deltas
are set to zero.  Along with the only possible event direction
GDK_SCROLL_SMOOTH it makes impossible to choose right direction
for zooming.  The issue has been fixed by disabling smooth events
if the cursor warping is set in the configuration.