lib: Rename make-pointer-to-pointer() and move it to (lepton ffi).
The new name reflects the fact that the procedure does the reverse
action compared with the stock Guile procedure dereference-pointer().

Moving it to the 'lepton' namespace will allow for dealing in
Scheme with other C code where pointers to pointers are used,
e.g. variables related to g_error_*() functions.