schematic: Do not use o_attrib_search_object_attribs_by_name() in Scheme code.
The function contents has been rewritten in Scheme and inlined in
the code of the &edit-slot() action.  Basically, it searches for
an attribute of a selected component by name, traversing first
its attached attribs and then its inherited ones.

Another obvious improvement is searching for the first component
in the selection list, not any primitive, as the C function
o_select_return_first_object() does.  Previously, the action
would silently quit when the first object was not a component.
Now, since the list of objects is filtered, the C function is no
longer necessary, either.