lib: Fix use of stat() in the (lepton file-system).
Calling stat() in regular-file?() and directory?()
functions with non-existent path raises an error
(exception).
Check whether given path exist before calling stat().

As a side effect, it fixes the behavior of the
component-library() function, which calls
add-component-library!(), which, in turn, calls
directory?(). If the latter throws an exception
(directory passed to component-library() does not
exist), processing of gafrc file stops and backtrace
is printed to the log.
It would be better to just ignore the wrong path
encountered, report it to the log and continue
reading gafrc.