symcheck: Fix random ITS test failures.
In some circumstances, lepton-symcheck integration tests might
fail due to two steps of reifying objects from C lists in
page-contents() and thus garbage collecting some of them in
between.  The first call of the procedure is in check-symbol(),
and the second one is just before producing a report.  As
check-symbol() already returns page objects, there is no point to
obtain them again from C lists.  As Scheme objects still exist and
none of them is garbage collected, their so named blaming info
cannot be lost, which might occur before as it is implemented as
'object-property' and hence a weak reference disappearing along
with its Scheme owner.  Rewriting the code in the way that
page-contents() is used once and check-report() retrieves objects
from check-symbol() itself thus protecting them from GC, makes the
code more functional and fixes the tests.