gui: Rewrite o_select_visible_unlocked() in Scheme.
The previous C code used o_attrib_add_selected() to select
attributes of every object on the current page.  The function ran
the hook 'select-objects-hook' for each set of attributes, which
was obviously wrong as in the end, o_select_visible_unlocked() ran
the same hook again for all objects on the page including their
attributes.  Now, the hook is called once for the resulting
selection list.