gui: Fix reporting an error when component is not found in library.
If the component library has changed and does not have the symbol
necessary for update a component, the user would see wrong message
in the log due to wrong argument #f in the function
set-component-with-transform().  This has been fixed by separating
code of making a new component and setting its properties.

Reported-by: graahnul-grom <graahnul.grom@gmail.com>