schematic: Get rid of several foreign functions in &hierarchy-down-schematic().
Apart from simplifying the code, the change fixes not so obvious
bug that was also present in the original C callback.  The code
wrongly incremented the 'count' variable even when the processing
of attached 'source=' attribs was done.  Instead of resetting the
counter to zero to start with the first inherited attribute, it
incremented it again.  In case when there were attached "source="
attributes though the files mentioned in them were missing or
unreadable, the code might try to deal with, say, the third
inherited 'source=' attribute instead of the first one.  This
might be if there were two attached 'source=' attributes.  In
a similar case, when the number of inherited 'source=' attributes
was less than the number of attached ones, the user might think
the program does the right thing by rejecting inherited attributes
if there were errors in dealing with attached ones.  However, it
was only a seeming behaviour.

The code has been amended to behave like lepton-netlist does in
such cases: if there are attached 'source=' attributes, only they
are processed; if the list of attached 'source=' attributes is
empty, the inherited attributes are processed.