schematic: Canonicalize symbol names to avoid opening new pages.
If a symbol name is relative, the code of hierarchy-down-symbol()
might not find an already open page for it since page names are
'normalized' on the C level when creating a page.  This issue has
been fixed by using Scheme function canonicalize-path() on all
file names returned by the function s_clib_symbol_get_filename().