sch2pcb: Split up the function parsing the 'schematics' project key.
A new function, sch2pcb_parse_schematics(), has been factored out
to separate two stages of processing the key.  On the first stage,
the key value is split up using the function g_shell_parse_argv()
and a list of new schematic names is formed.  On the second stage,
the list is added to the list of schematic names stored in the
global variable 'schematics'.