lepton: Amend determining home directory in expand-env-variable().
The function will now fall back to lower level functions to
determine the user home directory if the $HOME environment
variable is not set.  A new local function, user-home-dir(), has
been created just for this task.  It is not exported as yet though
may be exported in future if needed.