netlist: Enable loading Scheme scripts from %load-path.
Previously, scripts specified by the options '-m' and '-l' were
required to be absolute file names or file names in the current
directory.  It is no longer necessary and now the user can simply
set the basename of a file to be loaded, possibly along with its
relative path, and the name will be searched for first in the
current directory, and next in all paths set in the %load-path
variable.  Thus, the directories set as user or system data paths
and listed in the output of functions 'user-data-dir' and
'sys-data-dirs' can be used for these purposes.  The first
readable file is loaded and its name is reported to the log.  If
no file was found, this is reported to log.