its: Add syntax rule allowing for launching commands without quoting.
It transforms syntax macro argument identifiers into strings
making them appropriate for feeding to the system*() call.
The call of the form:

  (command-values* a "b" c)

transforms into

  (command-values "a" "b" "c")

which, in turn, launches system*() exact with these arguments.

Additionally, if an argument has prefix '$HOME/', its string is
expanded and the value of the "HOME" environment variable is
substituted for the prefix.