schematic: Delete undo backup files only when they exist.
There may be situtations when undo backup files don't exist when
the user tries to delete them.  The one I've triggered can be
reproduced with the following steps:

- Make some changes in schematic and then redo all them.
- Make some other changes and use 'F Q' to exit the program.

In such a case, the program used to report an error when the
function delete-file() was called.  This has been fixed by
checking if the backup file to delete exists.