schematic: Add actions and keystrokes for dealing with buffers.
The number of paste buffers is now 10 and the user can cut, copy,
and paste to/from any of them by prefixing the usual commands
Ctrl-X, Ctrl-C, and Ctrl-V with a numerical prefix from 0 to 9.
Zero prefix does the same that the keystroke without prefix does,
that is, works with clipboard buffer.