schematic: Factor out a new module, (schematic mouse-pointer)
As of now, the module contains an only function, pointer-position().

The module name reflects the fact that it deals with the mouse
pointer as the prefix 'pointer' is used in Guile to name FFI
functions working with C pointers.  The module name has been
chosen to avoid such an ambiguity.