sch2pcb: Avoid using fclose() on NULL pointers.
fclose() does not support NULL as its argument.  Moreover, if it
gets NULL as the argument, its behavior is not specified and it
may segfault on some systems, e.g. on FreeBSD.  Besides, if
fopen() returned NULL, it opened no file and there is nothing to
close so the call for fclose() is superfluous and can be safely
dropped in such cases.