lib: Fix possible unit test failure.
In the "component-library-command" unit test, the
code creates an sh script with tmpnam(), chmod()
it to 755, then this script is executed by the
make-component/library() function.

tmpnam() usually create files in /tmp.
If /tmp is on a separate file system, which is
mounted with the 'noexec' mount(8) option (common
practice), the script won't run, and the unit
test will fail.

Run that script with 'sh <script-name>' command.