gui: Enable creating orthogonal connections of nets to buses
By default, when 'bus-ripper-type' is set to 'net', during
creating a connection, a net connecting to a bus changes its
direction by 45 degrees and connects diagonally.  Sometimes the
user wants to make an orthogonal net-bus connections.  In this
commit, zero length bus rippers have been allowed in the
configuration to achieve this.