schematic: Add a new foreign structure, <canvas>
A new module, (schematic canvas foreign), has been added, which
contains helpers for working with wrapped pointers of this type.
The new Scheme structure and the module are intended to simplify
working with the foreign C type 'GschemPageView' and its
functions.  The name 'canvas' has been chosen deliberately in
order to reduce the function and variable names for the type.