gui: Avoid direct reference to dialog's parent structure
The parent SchematicWindow instance pointer for a SchematicNewText
instance can be absolutely equally obtained via its
SchematicDialog parent, or directly by using g_object_get() as
that function uses the same parent to retrieve that data.  As the
function is already used just above in the code to get the data,
the code has been simplified by removing redundant pointers.