schematic: Rename the function pointer-position() to mouse-pointer-position()
The function has been renamed to avoid ambiguities due to Guile's
use of the prefix 'pointer-' for functions working with foreign C
pointers.

A sibling of the function with the previous name has been added to
preserve backwards compatibility.  The docstring of the function
contains info on its deprecation.  A log warning about this is
issued as well during its evaluation.