gui: Fix a button sensitivity bug in Find text widget
If the widget entry contains any text, it is restored along with
the entry widget itself if the user first switches to 'Check
symbol' and then to 'Find text', for example.  However, the
sensitivity of the Find button is switched off anyway independent
of the text presence and until some other text is entered.  The
sensitivity of the button must depend on the presence of any text
in the entry.  An excerpt from another callback already
implementing such a behaviour has been re-used to fix the issue.