gui: Get rid of processing the LeptonToplevel structure when creating images
The structure is internally referred to by the parent
SchematicWindow structure so there is no point to copy and do
anything other with it as we don't want to get any values from it
in the function.

We just pull out one variable from it, the active page, to get the
image we want.