schematic: Catch the "delete-event" signal to avoid destroying the Multiattrib dialog
This commit fixes the bug introduced in the previous one.  When
the user hits the "Close" button in the Multiattrib dialog, the
dialog gets hidden and is not destroyed.  OTOH, if the user hits
Escape or closes the window using a WM shortkey or its delete
button, the "delete-event" signal leads to destroying the parent
GtkDialog structure, which leads to crashes if the user tries to
open the dialog window once again.  The issue has been fixed by
catching the signal in a new callback and stopping propagating it
to avoid processing it in the lower level code of the GtkDialog
parent of the widget.