netlist: Fix the sequence of evaluations in the 'drc' backend
After transforming 'drc' into a module backend, running
'lepton-netlist --list-backends' started to fail due to early
evaluation of the code of one function that looks for the
'attribs' config file.  The issue could not be triggered when the
backend was legacy as the code of those is primitive loaded and
thus evaluated only at run time while for module backends, it is
evaluated during their compilation as well.  The issue has been
fixed by avoiding early evaluation of the culprit function in the
code.