netlist: Use vams backend options to set netlisting mode and attributes
This is a preparation for transforming the 'vams' backend from
legacy into the module one.  In the case of a module backend we
cannot use toplevel definitions to evaluate toplevel expressions
from inside the module.  Thus, the backend options are applied to
get the same functionality.

The tests have been amended accordingly to the change.