gui: Avoid zooming when zoom direction is undefined
The commit introduces a new zooming constant, 'ZOOM_SAME', which,
in case the direction of zooming cannot be determined, prevents
zooming.  For 'non-smooth' events, this avoids doing anything in
case of a programming error.  For 'smooth' events, for GTK3 port
only, this prevents choosing a wrong zoom direction at the
beginning of zooming.  Thus the change fixes a first event zooming
bug in the current GTK3 version of lepton-schematic when the first
'smooth' event always caused zooming in despite of the mouse wheel
rotation direction.