tools: Set XDG_CONFIG_DIRS to predefined value when running from the build dir
This setting allows to start lepton-schematic GUI from the build
subdirectory "lepton-eda".  However, as the program configuration
files are in another location, the menus and some other functions
are not available yet in the GUI.

The new config system path lookup function introduced in one of
the previous commits has been utilized in system-gafrc to support
the new configuration and make 'make distcheck' work.