build: Move symbol library autoload script to lepton-eda/
The config-symbol-libraries.scm script has been moved to a new
autoload directory residing in the 'lepton-eda' directory in the
root of the repository.  As the directory is in XDG_CONFIG_DIRS
when lepton-schematic runs without installation, the script is
loaded automatically the same way as for other installed code, and
the system symbols are always available for the user.