schematic: Fix icon names for both GTK2 and GTK3 ports
The icons named 'edit-undo' and 'edit-redo' seem to be unavailable
for GTK2 port.  This has been fixed by replacing them with 'undo'
and 'redo', correspondingly.  Replacing icons 'gtk-goto-bottom'
and 'gtk-go-down' with 'go-bottom' and 'go-down' makes them
displayed in the GTK2 port of lepton-schematic as well as improves
the look of the GTK3 port as the new icons are taken from the same
theme as other icons in the toolbar (at least on my system).  The
former pair looked differently than other stock icons.