sch2pcb: Sort files with duplicate names when looking up a file element
When a directory to search for a file element contains multiple
files with the given name, there is no way to predict the order
they're found as most of stock Scheme functions dealing with
directory streams like readdir() or file-system-fold() do not sort
paths when requested to return the next one.  To avoid ambiguity
in such cases, the paths found are now sorted using the function
sort-locale<?() and the first item of the resulting list is
returned.

The unit tests have been fixed accordingly.