schematic: Fix canvas callback crashes due to undefined (current-window)
Scheme callback code may call other Scheme functions that need
some info from the current window such as coordinates and so on.
This information may be unavailable as the callbacks are called
via C Glib signals and may be evaluated even in different threads.
In such cases, to avoid crashes due to undefined current-window(),
the guard with-window() is now used to properly set the necessary
window fluid within the callback scope.