lepton: Avoid sporadic test suite failures on log rotation
When several processes of a lepton tool, say lepton-netlist, are
running in parallel and producing many log files, they can also
try to delete same files while rotating logs.  It causes
misleading error messages like

  ERROR: Failed to load RC file [.../system-gafrc]: No such file or directory

as the log rotation starts from the RC script system-gafrc, and
the current RC handlers are not smart enough to show the real
culprit.

Moreover, the user is not always able to find an actual error even
in the log files as they might be deleted by another process that
started log rotation at the same time.

The implemented work-around for the issue is prohibiting the log
rotation when the environment variable LEPTON_INHIBIT_RC_FILES is
set, which is true during the tests.