gui: Fix displaying of several widgets in dialogs
Instead of dealing with pointers to pointers, setters are now used
to store created dialogs in appropriate fields of the parent
SchematicWindow instance.

The commit fixes the bug of displaying of several widgets in their
dialogs ('schematic.gui::use-docks' is set to 'false').  When the
widgets were shown the second time, an empty dialog appeared and
the program output a GTK warning like the following in the
terminal:

  Gtk-WARNING **: Attempting to add a widget with type
  SchematicLogWidget to a container of type GtkBox, but the widget
  is already inside a container of type GtkBox, please remove the
  widget from its existing container first.

The widgets that have been fixed are:
- SchematicOptionsWidget (Options -> Options... or 'oo')
- SchematicTextPropertiesWidget (Edit -> Text... or 'ex' on text)
- SchematicObjectPropertiesWidget (Edit -> Object Properties... or
  'eo' on graphic primitives)
- SchematicLogWidget (Options -> Log Window or 'ol)
- SchematicFindTextState (View -> Find Text Results...)
- ColorEditWidget (View -> Color Scheme Editor...)
- FontSelectWidget (Options -> Font...)
- PageSelectWidget (Page -> Manager...)